refactor: merge metadata.py into meta_parser.py
